refactor: move to contract package
