+++ presentation/src/main/kotlin/jp/co/yumemi/android/code_check/presentation/feature/search/top/contract/SearchTopEvent.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check.presentation.feature.search.top
+package jp.co.yumemi.android.code_check.presentation.feature.search.top.contract
 
 import jp.co.yumemi.android.code_check.domain.entities.SearchType
 import jp.co.yumemi.android.code_check.presentation.core.contract.Event
